# swagger.yaml
openapi: 3.0.3
info:
  title: Site e-commerce
  description: >
    Some useful links:

    - [Lien Github du code
    source](https://github.com/DavyHaYehoudi/e-commerce-Cindy)

    - [The source API definition for the Pet
    Store](https://github.com/swagger-api/swagger-petstore/blob/master/src/main/resources/openapi.yaml)
  contact:
    email: davy.programa@gmail.com
  version: 1.0.0
servers:
  - url: https://ateliernoralya.com
...
paths:
  /clients:
    get:
      tags:
        - client
      summary: Get all customers
      description: This route retrieves a list of all the customers registered in the system. Information about each customer, such as their name, address, e-mail address, etc., can be included in the response.Route reserved for administrators
      operationId: getAllClients
      parameters:
        - name: itemsPerPage
          in: query
          description: Number of items per page
          required: true
          schema:
            type: integer
            default: "5"
        - name: steps
          in: query
          description: Number of steps
          required: false
          schema:
            type: string
            default: ""
            example: 0,3,5
        - name: trackingNumber
          in: query
          description: trackingNumber in order 
          required: false
          schema:
            type: string
            default: ""
            example: true
        - name: note
          in: query
          description: note in client 
          required: false
          schema:
            type: string
            default: ""
            example: true
        - name: credit
          in: query
          description: credit in orderProducts 
          required: false
          schema:
            type: string
            default: ""
            example: true
        - name: exchange
          in: query
          description: exchange in orderProducts 
          required: false
          schema:
            type: string
            default: ""
            example: true
        - name: refund
          in: query
          description: refund in orderProducts 
          required: false
          schema:
            type: string
            default: ""
            example: true
        - name: preciseDate
          in: query
          description: precise creation date of order 
          required: false
          schema:
            type: string
            default: ""
            example: 2024-03-10 in this format
        - name: rangeDateStart
          in: query
          description: start of range creation date of order 
          required: false
          schema:
            type: string
            default: ""
            example: 2024-03-10 in this format
        - name: rangeDateEnd
          in: query
          description: end of range creation date of order 
          required: false
          schema:
            type: string
            default: ""
            example: 2024-05-20 in this format
      responses:
        '200':
          description: successful getting all clients
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Clients'
        '403':
          description: Unauthorized - Admin is not authenticated
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Accès refusé. Vous n'êtes pas un administrateur.
      security:
        - api_key: []
...
  /materials/{materialId}:
    patch:
      tags:
        - material
      summary: Update a material
      description: This route, reserved for administrators, is used to modify the details of a specific material identified by its unique identifier (materialId). Fields to be updated, such as the material's name and color, are provided in the PATCH request body.
      operationId: updateMaterial
      parameters:
        - name: materialId
          in: path
          description: ID of the material that needs to be updated
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MaterialEdit'
      responses:
        '200':
          description: successful updated material
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MaterialCreateResponse'
        '403':
          description: Unauthorized - Admin is not authenticated
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Accès refusé. Vous n'êtes pas un administrateur.
        '404':
          description: The ressource does not exist.
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Le matériau n'existe pas.
      security:
        - api_key: []
    delete:
      tags:
        - material
      summary: Delete a material
      description: This route, reserved for administrators, allows deleting a specific material from the list of available materials based on its unique identifier (materialId).
      operationId: deleteMaterial
      parameters:
        - name: materialId
          in: path
          description: ID of the material that needs to be deleted
          required: true
          schema:
            type: string
      responses:
        '200':
          description: successful deleted material
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MaterialCreateResponse'
        '403':
          description: Unauthorized - Admin is not authenticated
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Accès refusé. Vous n'êtes pas un administrateur.
        '404':
          description: The ressource does not exist.
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Le matériau n'existe pas.
      security:
        - api_key: []
...
  /collections/{collectionId}:
    put:
      tags:
        - collection
      summary: Update a collection
      description: This route, reserved for administrators, allows updating an existing collection. Details of the collection to be updated, such as its name and the products it contains, are provided in the PUT request body.
      operationId: updateCollection
      parameters:
        - name: collectionId
          in: path
          description: ID of the collection that needs to be updated
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CollectionEdit'
      responses:
        '200':
          description: Successful updated collection
          content:
           application/json:
            schema:
              $ref: '#/components/schemas/Collection'
        '403':
          description: Unauthorized - Admin is not authenticated
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Accès refusé. Vous n'êtes pas un administrateur.
        '404':
          description: The ressource does not exist.
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: La collection n'existe pas.
      security:
        - api_key: []
    delete:
      tags:
        - collection
      summary: Delete collection
      description: This route, reserved for administrators, allows deleting a specific collection.
      operationId: deleteCollection
      parameters:
        - name: collectionId
          in: path
          description: ID of the collection that needs to be deleted
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Successful deleted collection
          content:
           application/json:
            schema:
              $ref: '#/components/schemas/Collection'
        '403':
          description: Unauthorized - Admin is not authenticated
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Accès refusé. Vous n'êtes pas un administrateur.
        '404':
          description: The ressource does not exist.
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: La collection n'existe pas.
      security:
        - api_key: []
...
  /categories/{categoryId}:
    patch:
      tags:
        - category
      summary: Update a category
      description: This route, reserved for administrators, allows updating an existing category. Details of the category to be updated, such as its name, the collection it belongs to, and the products it contains, are provided in the PATCH request body
      operationId: updateCategory
      parameters:
        - name: categoryId
          in: path
          description: ID of the category that needs to be updated
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CategoryEdit'
      responses:
        '200':
          description: Successful updated category
          content:
           application/json:
            schema:
              $ref: '#/components/schemas/Category'
        '403':
          description: Unauthorized - Admin is not authenticated
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Accès refusé. Vous n'êtes pas un administrateur.
        '404':
          description: The ressource does not exist.
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: La catégorie n'existe pas.
      security:
        - api_key: []
    delete:
      tags:
        - category
      summary: Delete a category
      description: This route, reserved for administrators, allows deleting a specific category
      operationId: deleteCategory
      parameters:
        - name: categoryId
          in: path
          description: ID of the category that needs to be deleted
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Successful deleted category
          content:
           application/json:
            schema:
              $ref: '#/components/schemas/Category'
        '403':
          description: Unauthorized - Admin is not authenticated
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Accès refusé. Vous n'êtes pas un administrateur.
        '404':
          description: The ressource does not exist.
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: La catégorie n'existe pas.
      security:
        - api_key: []
...
components:
  schemas:
    Clients:
      type: array
      items:
        $ref: '#/components/schemas/ClientDetails'
      xml:
        name: clients
    ClientDetails:
      type: object
      properties:
        id:
          type: string
          example: 5fc4d22b45a0fb3fd4a75f65
        firstName:
          type: string
          example: John
        lastName:
          type: string
          example: Doe
        email:
          type: string
          example: john@example.com
        phone:
          type: string
          example: 06 90 12 34 56
        avatar:
          type: string
        role:
          type: string
          example: user
        shippingAddress:
          type: object
          properties:
            firstName:
              type: string
            lastName: 
              type: string
            street: 
              type: string
            apartment: 
              type: string
            city:
              type: string
            postalCode: 
              type: string
            country: 
              type: string
            email: 
              type: string
            phone: 
              type: string
        billingingAddress:
          type: object
          properties:
            companyName:
              type: string
            firstName:
              type: string
            lastName: 
              type: string
            street: 
              type: string
            apartment: 
              type: string
            city:
              type: string
            postalCode: 
              type: string
            country: 
              type: string
            email: 
              type: string
            phone: 
              type: string
        totalOrders:
          type: integer
          format: int16
          example: 3
        wishlist:
          type: array
          items:
            type: object
            properties:
              productsId:
                type: string
                example: 6vc4d22b45a0fb3fd4a75f33
              material:
                type: integer
                format: int8
                example: 1
        cart:
          type: array
          items:
            type: object
            properties:
              productsId:
                type: string
                example: 1jc4d22b45a0fb3fd4a75f50
              material:
                type: integer
                format: int8
                example: 3
        Orders:
          type: array
          items:
            type: string
          example:
            - 2884d22b45a0fb3fd4a75p2
            - 7774d22b45a0fb3fd4a7bn0
            - 10m4d22b45a0fb3fd4a7l53
...
    MaterialEdit:
      type: object
      properties:
        name:
          type: string
          example: Or rose
    MaterialCreateResponse:
      type: object
      properties:
        _id:
          type: string
          example: 2884d22b45a0fb3fd4a75p2
        name:
          type: string
          example: Or rose
        value:
          type: string
          example: "#F0F0F0"
...
    Collection:
      type: array
      items:
        type: object
        properties:
          _id:
            type: string
            example: 2884d22b45a0fb3fd4a75p2
          name:
            type: string
            example: Pour un jour unique
          main_image:
            type: string
...
    CollectionEdit:
      type: object
      properties:
        main_image:
          type: string
    Category:
      type: array
      items:
        type: object
        properties:
          _id:
            type: string
            example: 2884d22b45a0fb3fd4a75p2
          name:
            type: string
            example: Accessoire cheveux
          parentCollection:
            type: array
            items: 
              type: string
              example: 
                - "65f2d2b24db605d02020061c"
                - "65f2d2b24db605d02020061d"
...
    CategoryEdit:
      type: object
      properties:
        name:
          type: string
          example: Sport
...
  securitySchemes:
    jwt_auth:
      type: apiKey
      in: header
      name: Authorization
      description: JWT authorization header using the Bearer scheme.
    api_key:
      type: apiKey
      name: api_key
      in: header